requirements.txt
numpy==1.26.4
pandas==2.2.1
optuna==3.6.1
plotly==5.20.0
streamlit==1.32.2
python-binance==1.0.19
duckdb==0.10.1
pyarrow==15.0.2
joblib==1.3.2
pyyaml==6.0.1
ta==0.11.0
scipy==1.12.0
loguru==0.7.2
click==8.1.7
ccxt==4.2.90
pytest==9.0.2
python-dotenv==1.2.1
python-dateutil==2.9.0
